+++ src/components/MuteButton.tsx
@@ -1,13 +1,14 @@
 import React, {useState} from "react";
 import {BasePlayerComponentButtonProps} from "../utils";
 import BaseButton from "./BaseButton";
-import {usePrestoCoreEvent} from "../react";
+import {usePrestoCoreEvent, usePrestoEnabledState} from "../react";
 
 export interface MuteButtonProps extends BasePlayerComponentButtonProps{
 }
 
 export const MuteButton = (props: MuteButtonProps) => {
   let [muted, setMuted] = useState(false);
+  let enabled = usePrestoEnabledState(props.player);
 
   usePrestoCoreEvent('volumechange', props.player, (e, presto) => {
     setMuted(presto.isMuted())
@@ -22,7 +23,7 @@ export const MuteButton = (props: MuteButtonProps) => {
   }
 
   return (
-    <BaseButton onClick={toggle} disableIcon={props.disableIcon}
+    <BaseButton onClick={toggle} disableIcon={props.disableIcon} disabled={!enabled}
                 className={`pp-ui-mute-toggle pp-ui-mute-toggle-${muted? "unmuted" : "muted"} ${props.className || ''}`}>
       {props.children}
     </BaseButton>