+++ src/components/RateText.tsx
@@ -1,13 +1,14 @@
 import React, {useState} from "react";
 import {BasePlayerComponentProps} from "../utils";
 import Label from "./Label";
-import {usePrestoUiEvent} from "../react";
+import {usePrestoEnabledStateClass, usePrestoUiEvent} from "../react";
 
 export interface RateTextProps extends BasePlayerComponentProps{
 }
 
 export const RateText = (props: RateTextProps) => {
   let [rate, setRate] = useState(1);
+  let enabledClass = usePrestoEnabledStateClass(props.player);
 
   usePrestoUiEvent('ratechange', props.player, (rate) => {
     if (rate != 0) {
@@ -16,7 +17,7 @@ export const RateText = (props: RateTextProps) => {
   })
 
   return (
-    <Label label={`x${rate}`} children={props.children} className={`pp-ui-label-rate ${props.className || ''}`}/>
+    <Label label={`x${rate}`} children={props.children} className={`pp-ui-label-rate ${enabledClass} ${props.className || ''}`}/>
   );
 }