+++ src/components/SeekButton.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import {BasePlayerComponentButtonProps} from "../utils";
 import BaseButton from "./BaseButton";
+import {usePrestoEnabledState} from "../react";
 
 export interface SeekButtonProps extends BasePlayerComponentButtonProps {
   seconds: number
@@ -8,11 +9,12 @@ export interface SeekButtonProps extends BasePlayerComponentButtonProps {
 
 export const SeekButton = (props: SeekButtonProps) => {
   const seek = () => props.player.position += props.seconds
+  let enabled = usePrestoEnabledState(props.player);
 
   if(!props.seconds) return <></>
 
   return (
-    <BaseButton onClick={seek} disableIcon={props.disableIcon}
+    <BaseButton onClick={seek} disableIcon={props.disableIcon} disabled={!enabled}
                 className={`pp-ui-seek pp-ui-seek-${props.seconds < 0 ? "back" : "forward"} ${props.className || ''}`}>
       {props.children}
     </BaseButton>